notification: make PluginTenantConfigurableConfigurationHandler#getConfigurable thread-safe
Under high load upon startup, getConfigurable was initializing the configuration for a tenant
multiple times. While the underlying map perTenantConfigurable is thread-safe, createConfigurable
was called by concurrent threads unnecessarily and the value re-initialized.

This could cause unexpected behaviors because when the old value is replaced, close() is called
on the configuration object while some threads might already been using it (note that this could
still happen when updating a configuration live, but this should be infrequent).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>