use Java 8 stream in PluginProperties. read more in next paragraph
When PluginProperties.buildPluginProperties() called, there's no guarantee that the first added map entry would be the first element of the List.

When using guava's ImmutableMap, it's known that guava make its implementation "sorted", so this assertion will always valid. This will not the same with JDK map. For short:

- "pluginProperties1.get(0).getKey()" using Guava's ImmutableMap.of() : Will always return "foo"
- "pluginProperties1.get(0).getKey()" using JDK Map.of() : No guarantee.

If someone wants to see old assertions, the commit hash is: 25abb935